Fix up numbers
(cherry picked from commit 8e2b551452a54e81fcaccc84e05acf2b632f4e1e)

+++ ComputerysBitStream/Core/ReadContext/ReadContext.State.cs
@@ -58,7 +58,7 @@ namespace ComputerysBitStream {
         /// <param name="capacity">New active bit capacity.</param>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public void SetCapacity(long capacity) {
-            ThrowIfInvalidCapacity(capacity, Buffer.Length * BitHelper.ULongSize);
+            ThrowIfInvalidCapacity(capacity, (long)Buffer.Length * BitHelper.ULongSize);
             ThrowIfInvalidPosition(Position, capacity);
             SetCapacityPrimitive(capacity);
         }
